Login Register Changpassword

+++ src/main/java/com/transport/transport/service/CRUDService.java
@@ -1,15 +1,17 @@
 package com.transport.transport.service;
 
+import com.transport.transport.model.entity.Account;
+
 import java.util.List;
 
 public interface CRUDService<T> {
     List<T> findAll();
 
-    T findById(int id);
+    T findById(Long id);
 
     void save(T entity);
 
-    void delete(int id);
+    void delete(Long id);
 
     void update(T entity);
 }
